backend/requirements.txt
faster-whisper==1.2.1
numpy==2.4.3
openai==2.29.0
Pillow==12.1.1
python-dotenv==1.2.2
pystray==0.19.5
pyttsx3==2.99
pywin32==311
RapidFuzz==3.14.3
sounddevice==0.5.5
